# Тили-тили/Тили-тили_API_openapi.yaml
openapi: 3.0.3
info:
  title: Тили-тили API
  version: 0.28.0
  description: |
    Контракт бэкенда для PWA «Тили-тили» (tili-tili.ru) — свадебный планировщик.
    Фронтенд (React) уже работает на моках; этот файл — единственный источник правды для бэкенда (Claude Code).

    ## Решения владельца 2026-09-02 (контракт v0.2)
    - Деньги — объект `{ amount: копейки, currency: ISO 4217 }` у КАЖДОЙ суммы.
      Не одна валюта на свадьбу: план §19.7 описывает свадьбу за границей, где
      сделка в евро, а бюджет в рублях. В MVP принимается только RUB.
    - У слота нет собственного статуса. Слот либо пуст, либо несёт сделку;
      состояние живёт в `Deal.state` — шесть значений. `Slot.tileState` —
      производная подпись для мозаики, только для чтения.
    - Гость опознаётся ОДНИМ персональным токеном во всех гостевых путях
      (`/rsvp/{guestToken}`, `/gifts/{guestToken}`, `/join/{guestToken}/…`).
      Общий код свадьбы не годится: по нему нельзя подставить имя в приглашение
      и нельзя дать гостю снять СВОЙ резерв подарка. Анонимность сохраняется —
      система знает гостя, а API пары этот токен не отдаёт никогда (§9).
    - Переноса данных с localStorage-моков не будет: реальных пользователей не
      было, эндпоинт импорта не заводится.

    ## Роли
    - `couple` — молодожёны (владельцы свадьбы): полный доступ.
    - `helper` — помощник (друг/родственник): орг-разделы, без бюджета и сделок.
...
    - `vendor` — подрядчик: свои сделки, лиды, чаты, анкета.
...

...
servers:
...
paths:
...
  /weddings:
    get:
      tags: [weddings]
      summary: Мои свадьбы
      description: |
        Единственный способ найти свою свадьбу после переустановки приложения:
        идентификатор хранится только на устройстве, и с чистым localStorage
        восстановить его больше неоткуда. Отдаёт и те свадьбы, куда человека
        позвали помощником или координатором, — с его ролью в каждой.
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  allOf:
                    - $ref: '#/components/schemas/Wedding'
                    - type: object
                      properties:
                        role: { type: string, enum: [couple, helper, coordinator, vendor] }
    post:
      tags: [weddings]
      summary: Создать свадьбу (после квиза)
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [partnerName, city]
              properties:
                partnerName: { type: string, example: Тимур }
                date: { type: string, format: date, example: '2027-06-14' }
                city: { $ref: '#/components/schemas/CityRef' }
                budgetTotal: { $ref: '#/components/schemas/Money' }  # example: 1500000
                guestsPlanned: { type: integer, example: 80 }
                style: { type: string, example: нежная классика }
                quizAnswers:
                  type: object
                  additionalProperties: true
      responses:
        '201':
          description: Свадьба создана; создатель — `couple`
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Wedding' }
  /weddings/{weddingId}:
    get:
      tags: [weddings]
      summary: Карточка свадьбы (данные для Home)
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/WeddingId' }]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Wedding' }
    patch:
      tags: [weddings]
      summary: Обновить (дата, город, бюджет, стиль)
      description: |
        Правка `date` — это ПЕРЕНОС: она проходит тот же путь, что
        `POST /weddings/{weddingId}/reschedule`, и тянет за собой занятость
        в календарях подрядчиков, сроки задач чек-листа, время блоков тайминга
        и час открытия чата дня X. Если кто-то из забронированной команды занят
        на новую дату — 409 `team_busy`, и не меняется ничего.

        Дата свадьбы принимается в диапазоне «год назад … пять лет вперёд»
        (решение владельца 2026-09-03); за его пределами — 422
        `date_out_of_range`. Несуществующий день вроде 30 февраля — 422 `bad_date`.
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/WeddingId' }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                date: { type: string, format: date }
                city: { $ref: '#/components/schemas/CityRef' }
                budgetTotal: { $ref: '#/components/schemas/Money' }
                guestsPlanned: { type: integer }
                style: { type: string }
                venue: { type: string, maxLength: 200 }
                tz: { type: string, maxLength: 64, description: 'явный пояс важнее пояса города' }
                inviteText: { type: string, maxLength: 2000, description: 'обращение пары в приглашении — его видит гость' }
                inviteThemeId:
                  type: integer
                  minimum: 0
                  maximum: 9
                  description: |
                    Сценарий оформления приглашения. Хранится у свадьбы, а не в
                    браузере пары: гость открывает приглашение со своего
                    устройства, и тема должна приехать к нему вместе с текстом.
                dressCode:
                  type: string
                  nullable: true
                  maxLength: 32
                  description: 'идентификатор палитры дресс-кода (План ч. 295) — по той же причине, что и тема; null — снять дресс-код'
                dressNote: { type: string, nullable: true, maxLength: 300, description: 'пожелание словами: «дамы — без белого»; null — убрать пожелание' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Wedding' }
        '409': { $ref: '#/components/responses/Conflict' }
  /weddings/{weddingId}/members:
    get:
      tags: [weddings]
      summary: Команда свадьбы
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/WeddingId' }]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Member' }
  /weddings/{weddingId}/members/{userId}:
    patch:
      tags: [weddings]
      summary: Сменить роль участника (только couple)
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/WeddingId' }
        - name: userId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [role]
              properties:
                role: { type: string, enum: [couple, helper, coordinator] }
      responses:
        '200': { description: OK }
        '409': { $ref: '#/components/responses/Conflict' }
    delete:
      tags: [weddings]
      summary: Удалить участника
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/WeddingId' }
        - name: userId
          in: path
          required: true
          schema: { type: string }
      responses:
        '204': { description: Удалён }
        '409': { $ref: '#/components/responses/Conflict' }

  # ───────────────────────── INVITES (команда) ─────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    WeddingId:
      name: weddingId
      in: path
      required: true
      schema: { type: string }
...
  responses:
...
    Conflict:
      description: Конфликт (уже существует)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code: { type: string }
            message: { type: string }
            fields:
              type: object
              additionalProperties: { type: string }
              description: |
                Только у 422: какие именно поля не прошли проверку.
                Ключ — имя поля (`reason`, `action`, `synonyms.тамада`),
                значение — что с ним не так. Человеку показывается `message`,
                поля подсвечивает форма.
...
    User:
      type: object
      properties:
        id: { type: string }
        name: { type: string, nullable: true, description: 'null, пока человек не назвал себя' }
        phone: { type: string, readOnly: true, description: 'E.164 — тот, что подтверждён кодом' }
        email: { type: string, format: email }
        avatarUrl: { type: string, nullable: true }
...
    CityRef:
      type: object
      required: [name, region]
      properties:
        name: { type: string }
        region: { type: string }
    Wedding:
      type: object
      properties:
        id: { type: string }
        title: { type: string, example: 'Алина ♥ Тимур' }
        date: { type: string, format: date }
        city: { $ref: '#/components/schemas/CityRef' }
        budgetTotal: { $ref: '#/components/schemas/Money' }
        guestsPlanned: { type: integer }
        style: { type: string }
        tz: { type: string, example: Asia/Yekaterinburg, description: 'таймзона места свадьбы. По ней открывается чат дня X и считаются напоминания — не по таймзоне пользователя' }
        inviteText: { type: string }
        inviteThemeId: { type: integer }
        dressCode: { type: string, nullable: true, description: 'палитра дресс-кода — её же видит гость' }
        dressNote: { type: string, nullable: true }
        cancelRequestedBy:
          type: string
          nullable: true
          description: >
            Кто из пары запросил отмену свадьбы и ждёт подтверждения второго.
            Видно только паре. Нужно, чтобы второй партнёр понимал: его нажатие
            «Отменить» не запросит отмену, а ИСПОЛНИТ её — брони отменятся,
            даты уйдут подрядчикам. Без этого поля показать предупреждение
            нечем. Пусто — запроса нет или он протух (срок 72 часа).
        cancelRequestedAt:
          type: string
          format: date-time
          nullable: true
          description: 'Когда запрошена отмена. Запрос живёт 72 часа, дальше требуется новый.'
        members:
          type: array
          items: { $ref: '#/components/schemas/Member' }
...
    Member:
      type: object
      description: >
        Член свадьбы — тот, у кого есть аккаунт и доступ в приложение.
        Гость (guest) и свой подрядчик (guest-vendor) членами НЕ являются:
        они опознаются токеном по ссылке, аккаунта не имеют и в members не попадают.
        Матрица доступа (§4 плана) описывает все шесть ролей, эта схема — только четыре с аккаунтом.
      properties:
        user: { $ref: '#/components/schemas/User' }
        role: { type: string, enum: [couple, helper, coordinator, vendor] }
        joinedAt: { type: string, format: date-time }
...
    Money:
      type: object
      description: |
        Денежная сумма. Решение владельца 2026-09-02: минорные единицы плюс код
        валюты у КАЖДОЙ суммы, а не одна валюта на свадьбу — план §19.7 описывает
        свадьбу за границей, где сделка в евро, а бюджет в рублях.
        В MVP принимается только RUB; enum расширяется вместе с поддержкой курсов.
      required: [amount, currency]
      properties:
        amount:
          type: integer
          minimum: 0
          maximum: 9007199254740991
          description: |
            Сумма в копейках. 85 000 ₽ = 8500000. Дробных значений не бывает.
            Потолок — предел точного целого в JSON: за ним число молча
            округляется, и «потрачено» перестаёт сходиться с суммой сделок.
          example: 8500000
        currency:
          type: string
          description: Код валюты по ISO 4217.
          enum: [RUB]
          example: RUB
...
security:
  - bearerAuth: []